+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- `--flat` option to emit the legacy fully unrolled address list.
+
+### Changed
+
+- Array instances now generate indexed macros/functions by default instead of enumerating every element.
+
 ## [0.1.3] - 2025-12-08
 
 ### Fixed